Move WeekDays components to the separate file
