+++ mobile/src/calendar/calendar-page.tsx
@@ -6,6 +6,7 @@ import { calendarStyles, calendarPeriodStyles, calendarPeriodColors } from './st
 import { DayModel, WeekModel, PeriodsModel, PeriodModel } from '../reducers/calendar/calendar.model';
 import { StartPeriod, EndPeriod, Period, DotPeriod } from './calendar-page-period';
 import { CalendarEventsType } from '../reducers/calendar/calendar-events.model';
+import { WeekDay, WeekDayCircle } from './calendar-page-weekday';
 
 export type OnSelectedDayCallback = (day: DayModel) => void;
 
@@ -109,9 +110,7 @@ export class CalendarPage extends Component<CalendarPageDefaultProps & CalendarP
             return null;
         }
 
-        const key = PeriodsModel.generateKey(date);
-
-        const periods = this.props.periods.get(key);
+        const periods = this.props.periods.get(date);
 
         if (!periods) {
             return null;
@@ -135,70 +134,5 @@ export class CalendarPage extends Component<CalendarPageDefaultProps & CalendarP
             default:
                 return null;
         }
-    }
-}
-
-export const WeekDay = (props: { hide: boolean, children: any[] }) =>
-    props.hide
-        ? null
-        : <View style={calendarStyles.weekDay}>{props.children}</View>;
-
-interface WeekDayCircleProps {
-    weekHeight: number;
-    day: DayModel;
-    onSelectedDay: OnSelectedDayCallback;
-}
-
-export class WeekDayCircle extends Component<WeekDayCircleProps> {
-    public render() {
-        const { day, weekHeight } = this.props;
-
-        const circleStyles = StyleSheet.flatten([
-            calendarStyles.weekDayCircle,
-            {
-                width: weekHeight,
-                height: weekHeight,
-                borderRadius: weekHeight / 2,
-                borderWidth: 2,
-                borderColor: day.today ? '#2FAFCC' : 'transparent',
-                backgroundColor: day.today
-                    ? '#fff'
-                    : 'transparent'
-            }
-        ]);
-
-        const innerCircleSize = (circleStyles.width as number) - (circleStyles.width as number * 0.2);
-
-        const innerCircleStyles = StyleSheet.flatten([
-            calendarStyles.weekDayCircle,
-            {
-                width: innerCircleSize,
-                height: innerCircleSize,
-                borderRadius: circleStyles.borderRadius - (circleStyles.borderRadius * 0.2),
-                backgroundColor: day.today
-                    ? '#2FAFCC'
-                    : 'transparent'
-            }
-        ]);
-
-        const circleTextStyles = StyleSheet.flatten({
-            color: day.belongsToCurrentMonth
-                ? day.today
-                    ? '#fff'
-                    : '#000'
-                : '#dadada'
-        });
-
-        return (
-            <TouchableHighlight style={circleStyles} onPress={this.onSelectedDay}>
-                <View style={innerCircleStyles}>
-                    <StyledText style={circleTextStyles}>{day.date.date()}</StyledText>
-                </View>
-            </TouchableHighlight>
-        );
-    }
-
-    private onSelectedDay = () => {
-        this.props.onSelectedDay(this.props.day);
     }
 }
